Rename participants_container
