+++ tests/models/input/container/test_participants_container.py
@@ -3,7 +3,7 @@ import os
 import pytest
 
 from definitions import ROOT_DIR
-from psdm_analysis.models.input.container.participants_container import (
+from psdm_analysis.models.input.container.participants import (
     SystemParticipantsContainer,
 )
